requirements.txt
discord.py>=2.6,<3.0
aiohttp>=3.9,<4.0
python-dotenv>=1.0,<2.0
aiosqlite>=0.20,<1.0
pytest>=8.0
pytest-asyncio>=0.23
pytest-cov>=5.0
